Implement better parsing of ptp4lconf
Previously, uds_address and message_tag were added at the end of the
conf passed to ptp4pl, based off of the incorrect assumption that this
was always the EOF.  This commit allows adding them to the global
section directly.
Also, this commit adds pulling in interfaces from ptp4lconf to use as
interface/message_tag, and moved interface to be passed through file,
instead of as a command line argument.