Add JUnit custom configurations
This commit adds a non-default junit report to the
serial and parallel e2e tests in order to produce
clean and easy-to-read report.

The default value of the junit flag was changed to be empty
because we want the junit report not to be produced
if the -junit flag is not present.

Signed-off-by: liornoy <lnoy@redhat.com>