Add k8s reporter to e2e tests
This commit utilizes the openshift-kni k8sreporter, and
pass it the CRDs, APIs and namespace of ptp.

The reporter will function only if the `-report` flag is passed.

This is to be aligned with the current reporter we use in the cnf-tests.

Also added a strings util to clean the test name so it can
be used to create a directory for the k8s report.

Signed-off-by: liornoy <lnoy@redhat.com>